requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchaudio==2.4.1+cu121

numpy==1.26.4
pillow==10.4.0
tqdm==4.66.5
matplotlib==3.9.2
